Built int-parser, refactored array()
Modifed the output of integer() so it returns an int instead of a string, further modified array() so it now returns a list of ints, BUT with the index now offset by the full length of s, rather than the number of elements in the int list. Also modified unittests accordingly.